api/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
httpx>=0.26.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Auth
clerk-backend-api>=0.2.0
PyJWT>=2.8.0

# Stripe
stripe>=7.0.0

# LLM Providers
openai>=1.10.0
anthropic>=0.18.0
google-generativeai>=0.3.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Utils
python-multipart>=0.0.6
